# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Moneyball Scout API",
  "description": "Live football data from API-Football, match ratings, form, valuations and manager ratings.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/teams/{team_id}": {
   "get": {
    "tags": [
     "teams"
    ],
    "summary": "Team Detail",
    "operationId": "team_detail_api_teams__team_id__get",
    "parameters": [
     {
      "name": "team_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Team Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TeamDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/managers/{coach_id}": {
   "get": {
    "tags": [
     "teams"
    ],
    "summary": "Manager Detail",
    "operationId": "manager_detail_api_managers__coach_id__get",
    "parameters": [
     {
      "name": "coach_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Coach Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ManagerDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ManagerDetail": {
    "properties": {
     "manager": {
      "$ref": "#/components/schemas/ManagerRow"
     },
     "nationality": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nationality"
     },
     "series": {
      "items": {
       "$ref": "#/components/schemas/ManagerPoint"
      },
      "type": "array",
      "title": "Series"
     }
    },
    "type": "object",
    "required": [
     "manager",
     "series"
    ],
    "title": "ManagerDetail"
   },
   "ManagerPoint": {
    "properties": {
     "date": {
      "type": "string",
      "format": "date",
      "title": "Date"
     },
     "pts": {
      "type": "integer",
      "title": "Pts"
     },
     "xpts": {
      "type": "number",
      "title": "Xpts"
     }
    },
    "type": "object",
    "required": [
     "date",
     "pts",
     "xpts"
    ],
    "title": "ManagerPoint"
   },
   "ManagerRow": {
    "properties": {
     "coach_id": {
      "type": "integer",
      "title": "Coach Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "photo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Photo"
     },
     "team": {
      "$ref": "#/components/schemas/TeamRef"
     },
     "tenure_start": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tenure Start"
     },
     "matches": {
      "type": "integer",
      "title": "Matches"
     },
     "ppg": {
      "type": "number",
      "title": "Ppg"
     },
     "xppg": {
      "type": "number",
      "title": "Xppg"
     },
     "pts_over_exp": {
      "type": "number",
      "title": "Pts Over Exp"
     },
     "squad_efficiency": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Squad Efficiency"
     },
     "development": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Development"
     },
     "score": {
      "type": "number",
      "title": "Score"
     }
    },
    "type": "object",
    "required": [
     "coach_id",
     "name",
     "team",
     "matches",
     "ppg",
     "xppg",
     "pts_over_exp",
     "score"
    ],
    "title": "ManagerRow"
   },
...
   "PlayerRow": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "photo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Photo"
     },
     "team": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/TeamRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "position_group": {
      "type": "string",
      "title": "Position Group"
     },
     "age": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Age"
     },
     "apps": {
      "type": "integer",
      "title": "Apps"
     },
     "minutes": {
      "type": "integer",
      "title": "Minutes"
     },
     "goals": {
      "type": "integer",
      "title": "Goals"
     },
     "assists": {
      "type": "integer",
      "title": "Assists"
     },
     "avg_rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Rating"
     },
     "form": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Form"
     },
     "expected_next": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expected Next"
     },
     "perf_index": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Perf Index"
     },
     "market_value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Market Value"
     },
     "p50": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P50"
     },
     "undervalue_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Undervalue Score"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "position_group",
     "apps",
     "minutes",
     "goals",
     "assists"
    ],
    "title": "PlayerRow"
   },
...
   "TeamDetail": {
    "properties": {
     "team": {
      "$ref": "#/components/schemas/TeamRow"
     },
     "series": {
      "items": {
       "$ref": "#/components/schemas/TeamMatchPoint"
      },
      "type": "array",
      "title": "Series"
     },
     "squad": {
      "items": {
       "$ref": "#/components/schemas/PlayerRow"
      },
      "type": "array",
      "title": "Squad"
     },
     "managers": {
      "items": {
       "$ref": "#/components/schemas/ManagerRow"
      },
      "type": "array",
      "title": "Managers"
     }
    },
    "type": "object",
    "required": [
     "team",
     "series",
     "squad",
     "managers"
    ],
    "title": "TeamDetail"
   },
   "TeamMatchPoint": {
    "properties": {
     "date": {
      "type": "string",
      "format": "date",
      "title": "Date"
     },
     "fixture_id": {
      "type": "integer",
      "title": "Fixture Id"
     },
     "opponent_id": {
      "type": "integer",
      "title": "Opponent Id"
     },
     "gf": {
      "type": "integer",
      "title": "Gf"
     },
     "ga": {
      "type": "integer",
      "title": "Ga"
     },
     "pts": {
      "type": "integer",
      "title": "Pts"
     },
     "xpts": {
      "type": "number",
      "title": "Xpts"
     }
    },
    "type": "object",
    "required": [
     "date",
     "fixture_id",
     "opponent_id",
     "gf",
     "ga",
     "pts",
     "xpts"
    ],
    "title": "TeamMatchPoint"
   },
...
   "TeamRef": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "logo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Logo"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name"
    ],
    "title": "TeamRef"
   },
   "TeamRow": {
    "properties": {
     "team": {
      "$ref": "#/components/schemas/TeamRef"
     },
     "attack": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Attack"
     },
     "defence": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Defence"
     },
     "rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rating"
     },
     "played": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Played"
     },
     "points": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Points"
     },
     "xpts": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Xpts"
     },
     "goals_for": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Goals For"
     },
     "goals_against": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Goals Against"
     },
     "avg_player_rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Player Rating"
     },
     "squad_value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Squad Value"
     }
    },
    "type": "object",
    "required": [
     "team"
    ],
    "title": "TeamRow"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}